Add check for existence of csv filename

+++ VScraper.py
@@ -3,6 +3,7 @@ from time import sleep
 import csv
 import requests
 from bs4 import BeautifulSoup as bs
+import os.path
 
 images = ['.png', '.jpg', '.jpeg', '.gif']
 audio = ['.mp3', '.mp4']
@@ -15,34 +16,37 @@ def get_files():
     the user's specified local directory.
     """
 
-    csv = input("Enter the CSV file you want to read from: ")
-    
-    with open(csv, 'rb') as csvfile:
-        filereader = csv.reader(csvfile, delimiter=' ', quotechar='|')
-        for row in filereader:
-            print(', ').join(row)
+    csvfilename = input("Enter the CSV file you want to read from: ")
+    if os.path.isfile(csvfilename):
+        print("File, " "'" + csvfilename + "'", "exists. Ready to read CSV.")
+        
+        with open(csvfilename, 'rb') as csvfile:
+            filereader = csv.reader(csvfile, delimiter=' ', quotechar='|')
+            for row in filereader:
+                print(', ').join(row)
 
-    
-    while True: 
-        suffix = input("\nWhat type of file do you want to scrape? \nExamples: images, audio, text - ")
-        filepath = input("Specify a file path to save to: ")
+        while True: 
+            suffix = input("\nWhat type of file do you want to scrape? \nExamples: images, audio, text - ")
+            filepath = input("Specify a file path to save to: ")
 
-        if not url.startswith('http://') and not url.startswith('https://'):
-            url += 'http://'
+            if not url.startswith('http://') and not url.startswith('https://'):
+                url += 'http://'
 
-        response = requests.get(url, stream=True)            
-        soup = bs(response.text)
+            response = requests.get(url, stream=True)            
+            soup = bs(response.text)
 
-        list_of_links = [link.get('href') for link in soup.find_all('a') if suffix in str(link)]
+            list_of_links = [link.get('href') for link in soup.find_all('a') if suffix in str(link)]
 
-        for link in list_of_links:
-            file_name = link.rpartition('/')[-1]
-            urlretrieve(url.rsplit('/', 1)[0] + '/' + link, filepath + '\\' + file_name)
-            
-        print_message(list_of_links, suffix)
-        if not repeat(input("\nScrape from another URL? ")):
-            break
+            for link in list_of_links:
+                file_name = link.rpartition('/')[-1]
+                urlretrieve(url.rsplit('/', 1)[0] + '/' + link, filepath + '\\' + file_name)
+                
+            print_message(list_of_links, suffix)
+            if not repeat(input("\nScrape from another URL? ")):
+                break
 
+    else:
+        print("File, " "'" + csvfilename + "'", "does not exist.")
 
 def print_message(lst, suffix):
     """ Notifies user when done downloading files OR
